remove unused program halgui
It was added in 2005 and never touched since.  It's never been installed
as part of our build system or packaging.  Our documentation has always
listed it as "not implemented yet".

It's the only part of our code that tries to import python-gnome2,
which is no longer included in Debian 10 ("Buster").

I'm removing it because i think no one uses it or cares about it.